+++ lib/lightning/policies/project_users.ex
@@ -74,6 +74,7 @@ defmodule Lightning.Policies.ProjectUsers do
              :edit_data_retention,
              :add_project_user,
              :remove_project_user,
+             :edit_run_settings,
              :create_collection
            ],
       do: project_user.role in [:owner, :admin]