+++ lib/lightning/runs/run_options.ex
@@ -22,6 +22,7 @@ defmodule Lightning.Runs.RunOptions do
     field :save_dataclips, :boolean, default: true
     field :run_timeout_ms, :integer, default: 60_000
     field :run_memory_limit_mb, :integer
+    field :enable_job_logs, :boolean
   end
 
   def new(opts \\ %{}) do
@@ -32,7 +33,12 @@ defmodule Lightning.Runs.RunOptions do
   defimpl Jason.Encoder, for: __MODULE__ do
     def encode(value, opts) do
       value
-      |> Map.take([:save_dataclips, :run_timeout_ms, :run_memory_limit_mb])
+      |> Map.take([
+        :save_dataclips,
+        :run_timeout_ms,
+        :run_memory_limit_mb,
+        :enable_job_logs
+      ])
       |> Map.reject(fn {_key, val} -> is_nil(val) end)
       |> Jason.Encode.map(opts)
     end