+++ priv/repo/migrations/20250226084032_add_enable_job_jobs_to_workflows_table.exs
@@ -1,0 +1,9 @@
+defmodule Lightning.Repo.Migrations.AddEnableJobJobsToWorkflowsTable do
+  use Ecto.Migration
+
+  def change do
+    alter table("workflows") do
+      add_if_not_exists :enable_job_logs, :boolean, default: true
+    end
+  end
+end